--configure option generates .rspec file instead of autotest/discover.rb

+++ lib/rspec/core/command_line_configuration.rb
@@ -16,7 +16,7 @@ module RSpec
 "#{@command}" is not valid a valid argument to "rspec --configure".
 Supported options are:
 
-  rspec --configure autotest # generates configuration to run autotest with rspec
+  rspec --configure autotest # generates .rspec file
 
 #{"*"*50}
 MESSAGE
@@ -26,26 +26,26 @@ MESSAGE
       class Autotest
         class << self
           def generate
-            create_autotest_directory
-            create_discover_file
-            puts "autotest/discover.rb has been added"
+            create_dot_rspec_file
+            remove_autotest_dir_if_present
           end
 
-          def create_autotest_directory
-            Dir.mkdir('autotest') unless File.exist?('autotest')
-          end
-
-          def create_discover_file
-            optionally_remove_discover_file if discover_file_exists?
-            File.open(discover_file_path, 'w') do |file|
-              file << 'Autotest.add_discovery { "rspec2" }'
+          def create_dot_rspec_file
+            puts "Autotest loads RSpec's Autotest subclass when there is a .rspec file in the project's root directory."
+            if File.exist?('./.rspec')
+              puts ".rspec file already exists, so nothing was changed."
+            else
+              FileUtils.touch('./.rspec')
+              puts ".rspec file did not exist, so it was created."
             end
           end
 
-          def optionally_remove_discover_file
-            print "Discover file already exists, overwrite [y/N]? "
-            exit if gets !~ /y/i
-            FileUtils.rm_rf(discover_file_path)
+          def remove_autotest_dir_if_present
+            if discover_file_exists?
+              print "Delete obsolete autotest/discover.rb [y/n]? "
+              exit if gets !~ /y/i
+              FileUtils.rm_rf(discover_file_path)
+            end
           end
 
           def discover_file_exists?
